+++ core/docs/changelog.md
@@ -11,6 +11,7 @@
 ### Changes
 * Update files expansion logic to use unified expansion method from container-engine-lib
 * Upgrade to container-engine-lib 0.27.0, which implements the unified file artifact expansion logic
+* `StartService` and `GetServiceInfo` now return the Service GUID
 
 # 1.50.0
 ### Features